Create FizzBuzz.js
Write a program that uses console.log to print all the numbers from 1 to 100, with exceptions. For numbers divisible by 3, print "Fizz" instead of the number, for numbers divisible by 5, print "Buzz" instead, and print "FizzBuzz" for numbers that are divisible by both 3 and 5 (and still print "Fizz" or "Buzz" for numbers divisible by only one of those).